+++ edgraph/access.go
@@ -75,3 +75,8 @@ func AuthorizeGuardians(ctx context.Context) error {
 	// always allow access
 	return nil
 }
+
+func AuthGuardianOfTheGalaxy(ctx context.Context) error {
+	// always allow access
+	return nil
+}
